fix results data and func
