+++ js/templates/game-one.js
@@ -7,7 +7,6 @@ import footer from './footer';
 import getOption from './options/option';
 import getScreen from '../game-switcher';
 import timer from '../helpers/timer';
-import getResult from './result';
 
 
 const createScreen = (data, gameStatistics, allAnswers) => {
@@ -91,7 +90,7 @@ const createScreen = (data, gameStatistics, allAnswers) => {
       if (lives > 0) {
         getScreen(++screenNumber, {time, lives, screenNumber, answers});
       } else {
-        addElementToPage(getResult({time, lives, screenNumber, answers}));
+        getScreen(null, {lives, answers});
       }
 
       deleteTimer();
